# perf: improve isAsyncFunction