--- a/lib/helpers/isAsyncFunction.js
+++ b/lib/helpers/isAsyncFunction.js
@@ -2,10 +2,13 @@
 
 const { inspect } = require('util');
 
-module.exports = function isAsyncFunction(v) {
-  if (typeof v !== 'function') {
-    return;
-  }
+const isAsyncFunctionCache = new WeakMap();
 
-  return inspect(v).startsWith('[AsyncFunction:');
+module.exports = function isAsyncFunction(v) {
+  if (isAsyncFunctionCache.has(v)) {
+    return isAsyncFunctionCache.get(v);
+  }
+  const result = typeof v === 'function' && inspect(v).indexOf('[AsyncFunction:') === 0 || false;
+  isAsyncFunctionCache.set(v, result);
+  return result;
 };