--- a/lib/helpers/isAsyncFunction.js
+++ b/lib/helpers/isAsyncFunction.js
@@ -2,9 +2,13 @@
 
 const { inspect } = require('util');
 
+const primitiveTypes = new Set(['number', 'string', 'bigint', 'boolean', 'symbol']);
 const isAsyncFunctionCache = new WeakMap();
 
 module.exports = function isAsyncFunction(v) {
+  if (primitiveTypes.has(v)) {
+    return false;
+  }
   if (isAsyncFunctionCache.has(v)) {
     return isAsyncFunctionCache.get(v);
   }