--- a/lib/helpers/isAsyncFunction.js
+++ b/lib/helpers/isAsyncFunction.js
@@ -6,7 +6,7 @@ const primitiveTypes = new Set(['number', 'string', 'bigint', 'boolean', 'symbol
 const isAsyncFunctionCache = new WeakMap();
 
 module.exports = function isAsyncFunction(v) {
-  if (primitiveTypes.has(v)) {
+  if (primitiveTypes.has(typeof v)) {
     return false;
   }
   if (isAsyncFunctionCache.has(v)) {