--- a/lib/helpers/isAsyncFunction.js
+++ b/lib/helpers/isAsyncFunction.js
@@ -2,7 +2,7 @@
 
 const { inspect } = require('util');
 
-const primitiveTypes = new Set(['number', 'string', 'bigint', 'boolean', 'symbol']);
+const primitiveTypes = new Set(['undefined', 'number', 'string', 'bigint', 'boolean', 'symbol']);
 const isAsyncFunctionCache = new WeakMap();
 
 module.exports = function isAsyncFunction(v) {