--- a/lib/helpers/isAsyncFunction.js
+++ b/lib/helpers/isAsyncFunction.js
@@ -1,18 +1,11 @@
 'use strict';
 
-const { inspect } = require('util');
-
 const primitiveTypes = new Set(['undefined', 'number', 'string', 'bigint', 'boolean', 'symbol']);
-const isAsyncFunctionCache = new WeakMap();
+const asyncFunctionPrototype = Object.getPrototypeOf(async function() {});
 
 module.exports = function isAsyncFunction(v) {
   if (primitiveTypes.has(typeof v)) {
     return false;
   }
-  if (isAsyncFunctionCache.has(v)) {
-    return isAsyncFunctionCache.get(v);
-  }
-  const result = typeof v === 'function' && inspect(v).indexOf('[AsyncFunction:') === 0 || false;
-  isAsyncFunctionCache.set(v, result);
-  return result;
+  return Object.getPrototypeOf(v) === asyncFunctionPrototype;
 };