--- a/lib/helpers/isAsyncFunction.js
+++ b/lib/helpers/isAsyncFunction.js
@@ -1,11 +1,10 @@
 'use strict';
 
-const primitiveTypes = new Set(['undefined', 'number', 'string', 'bigint', 'boolean', 'symbol']);
 const asyncFunctionPrototype = Object.getPrototypeOf(async function() {});
 
 module.exports = function isAsyncFunction(v) {
-  if (primitiveTypes.has(typeof v)) {
-    return false;
-  }
-  return Object.getPrototypeOf(v) === asyncFunctionPrototype;
+  return (
+    typeof v === 'function' &&
+    Object.getPrototypeOf(v) === asyncFunctionPrototype
+  );
 };