@Uzlopak 2022-02-17
With the improved benchmark from #11415 
before: 

before:
```
MONGOOSE_DEV=1 node validate
invalid async x 21,179 ops/sec ±0.71% (84 runs sampled)
valid async x 41,408 ops/sec ±1.34% (83 runs sampled)
invalid sync x 24,553 ops/sec ±0.76% (93 runs sampled)
valid sync x 76,212 ops/sec ±0.67% (92 runs sampled)
```
after:
```
 MONGOOSE_DEV=1 node validate
invalid async x 21,100 ops/sec ±0.80% (83 runs sampled)
valid async x 42,095 ops/sec ±1.59% (81 runs sampled)
invalid sync x 27,730 ops/sec ±0.59% (93 runs sampled)
valid sync x 105,097 ops/sec ±0.68% (95 runs sampled)
```

As you can see it improves the valid sync path significantly.